+++ FirebaseAI/Sources/GenerativeAIService.swift
@@ -202,7 +202,7 @@ struct GenerativeAIService {
   private func logRPCError(_ error: BackendError) {
     let projectID = firebaseInfo.projectID
     if error.isVertexAIInFirebaseServiceDisabledError() {
-      AILog.error(code: .agentPlatformInFirebaseAPIDisabled, """
+      AILog.error(code: .enterpriseInFirebaseAPIDisabled, """
       The Firebase AI SDK requires the Firebase AI API \
       (`firebasevertexai.googleapis.com`) to be enabled in your Firebase project. Enable this API \
       by visiting the Firebase Console at