+++ FirebaseAI/Sources/Types/Public/Backend.swift
@@ -23,12 +23,12 @@ public struct Backend {
   /// locations](https://firebase.google.com/docs/ai-logic/locations?api=vertex#available-locations)
   /// for a list of supported locations.
   @available(*, deprecated, message: """
-  Use agentPlatform(location:) instead; note that the default location is now "global" instead of "us-central1"
+  Use enterprise(location:) instead; note that the default location is now "global" instead of "us-central1"
   """)
   public static func vertexAI() -> Backend {
     return Backend(
       apiConfig: APIConfig(
-        service: .agentPlatform(endpoint: .firebaseProxyProd, location: "us-central1"),
+        service: .enterprise(endpoint: .firebaseProxyProd, location: "us-central1"),
         version: .v1beta
       )
     )
@@ -41,31 +41,43 @@ public struct Backend {
   ///     for a list of supported locations.
   ///
   /// [1]: https://firebase.google.com/docs/ai-logic/locations?api=vertex#available-locations
-  @available(*, deprecated, renamed: "agentPlatform(location:)", message: """
-  Vertex AI has been renamed to the Agent Platform Gemini API.
+  @available(*, deprecated, renamed: "enterprise(location:)", message: """
+  Vertex AI has been renamed to the Gemini Enterprise API.
   """)
   public static func vertexAI(location: String) -> Backend {
     return Backend(
       apiConfig: APIConfig(
-        service: .agentPlatform(endpoint: .firebaseProxyProd, location: location),
+        service: .enterprise(endpoint: .firebaseProxyProd, location: location),
         version: .v1beta
       )
     )
   }
 
-  /// Initializes a `Backend` configured for the Agent Platform Gemini API.
+  /// Initializes a `Backend` configured for the Gemini Enterprise API.
   ///
-  /// > Note: The Agent Platform Gemini API was formerly known as Vertex AI.
+  /// > Note: The Gemini Enterprise API was formerly known as Vertex AI.
   ///
   /// - Parameters:
   ///   - location: The region identifier, defaulting to `global`; see
   ///     [available locations][1] for a list of supported locations.
   ///
   /// [1]: https://firebase.google.com/docs/ai-logic/locations?api=vertex#available-locations
+  public static func enterprise(location: String = "global") -> Backend {
+    return Backend(
+      apiConfig: APIConfig(
+        service: .enterprise(endpoint: .firebaseProxyProd, location: location),
+        version: .v1beta
+      )
+    )
+  }
+
+  @available(*, deprecated, renamed: "enterprise(location:)", message: """
+  The Agent Platform Gemini API has been renamed to the Gemini Enterprise API.
+  """)
   public static func agentPlatform(location: String = "global") -> Backend {
     return Backend(
       apiConfig: APIConfig(
-        service: .agentPlatform(endpoint: .firebaseProxyProd, location: location),
+        service: .enterprise(endpoint: .firebaseProxyProd, location: location),
         version: .v1beta
       )
     )