+++ FirebaseAI/Tests/TestApp/Tests/Integration/GenerateContentIntegrationTests.swift
@@ -48,9 +48,9 @@ struct GenerateContentIntegrationTests {
   }
 
   @Test(arguments: [
-    (InstanceConfig.agentPlatform_v1beta, ModelNames.gemini2_5_FlashLite),
-    (InstanceConfig.agentPlatform_v1beta_global, ModelNames.gemini2_5_FlashLite),
-    (InstanceConfig.agentPlatform_v1beta_global_appCheckLimitedUse, ModelNames.gemini2_5_FlashLite),
+    (InstanceConfig.enterprise_v1beta, ModelNames.gemini2_5_FlashLite),
+    (InstanceConfig.enterprise_v1beta_global, ModelNames.gemini2_5_FlashLite),
+    (InstanceConfig.enterprise_v1beta_global_appCheckLimitedUse, ModelNames.gemini2_5_FlashLite),
     (InstanceConfig.googleAI_v1beta, ModelNames.gemini3_1_FlashLite),
     (InstanceConfig.googleAI_v1beta_appCheckLimitedUse, ModelNames.gemini3_1_FlashLite),
     (InstanceConfig.googleAI_v1beta, ModelNames.gemma4_31B),
@@ -149,18 +149,18 @@ struct GenerateContentIntegrationTests {
 
   @Test(
     arguments: [
-      (.agentPlatform_v1beta, ModelNames.gemini2_5_Flash, ThinkingConfig(thinkingBudget: 0)),
-      (.agentPlatform_v1beta, ModelNames.gemini2_5_Flash, ThinkingConfig(thinkingBudget: 24576)),
-      (.agentPlatform_v1beta, ModelNames.gemini2_5_Flash, ThinkingConfig(
+      (.enterprise_v1beta, ModelNames.gemini2_5_Flash, ThinkingConfig(thinkingBudget: 0)),
+      (.enterprise_v1beta, ModelNames.gemini2_5_Flash, ThinkingConfig(thinkingBudget: 24576)),
+      (.enterprise_v1beta, ModelNames.gemini2_5_Flash, ThinkingConfig(
         thinkingBudget: 24576, includeThoughts: true
       )),
-      (.agentPlatform_v1beta_global, ModelNames.gemini2_5_Pro, ThinkingConfig(thinkingBudget: 128)),
+      (.enterprise_v1beta_global, ModelNames.gemini2_5_Pro, ThinkingConfig(thinkingBudget: 128)),
       (
-        .agentPlatform_v1beta_global,
+        .enterprise_v1beta_global,
         ModelNames.gemini2_5_Pro,
         ThinkingConfig(thinkingBudget: 32768)
       ),
-      (.agentPlatform_v1beta_global, ModelNames.gemini2_5_Pro, ThinkingConfig(
+      (.enterprise_v1beta_global, ModelNames.gemini2_5_Pro, ThinkingConfig(
         thinkingBudget: 32768, includeThoughts: true
       )),
       (.googleAI_v1beta, ModelNames.gemini2_5_FlashLite, ThinkingConfig(thinkingBudget: 0)),
@@ -272,15 +272,15 @@ struct GenerateContentIntegrationTests {
   @Test(
     arguments: [
       (
-        .agentPlatform_v1beta_global,
+        .enterprise_v1beta_global,
         ModelNames.gemini2_5_Flash,
         ThinkingConfig(thinkingBudget: -1)
       ),
-      (.agentPlatform_v1beta_global, ModelNames.gemini2_5_Flash, ThinkingConfig(
+      (.enterprise_v1beta_global, ModelNames.gemini2_5_Flash, ThinkingConfig(
         thinkingBudget: -1, includeThoughts: true
       )),
-      (.agentPlatform_v1beta_global, ModelNames.gemini2_5_Pro, ThinkingConfig(thinkingBudget: -1)),
-      (.agentPlatform_v1beta_global, ModelNames.gemini2_5_Pro, ThinkingConfig(
+      (.enterprise_v1beta_global, ModelNames.gemini2_5_Pro, ThinkingConfig(thinkingBudget: -1)),
+      (.enterprise_v1beta_global, ModelNames.gemini2_5_Pro, ThinkingConfig(
         thinkingBudget: -1, includeThoughts: true
       )),
       (.googleAI_v1beta, ModelNames.gemini2_5_FlashLite, ThinkingConfig(thinkingBudget: -1)),
@@ -358,11 +358,11 @@ struct GenerateContentIntegrationTests {
   }
 
   @Test(arguments: [
-    (InstanceConfig.agentPlatform_v1beta, ModelNames.gemini2_5_FlashImage),
-    (InstanceConfig.agentPlatform_v1beta_global, ModelNames.gemini2_5_FlashImage),
+    (InstanceConfig.enterprise_v1beta, ModelNames.gemini2_5_FlashImage),
+    (InstanceConfig.enterprise_v1beta_global, ModelNames.gemini2_5_FlashImage),
     (InstanceConfig.googleAI_v1beta, ModelNames.gemini2_5_FlashImage),
     (InstanceConfig.googleAI_v1beta, ModelNames.gemini3_1_FlashImage),
-    (InstanceConfig.agentPlatform_v1beta_global, ModelNames.gemini3_1_FlashImage),
+    (InstanceConfig.enterprise_v1beta_global, ModelNames.gemini3_1_FlashImage),
   ])
   func generateImageWithAspectRatio(_ config: InstanceConfig, modelName: String) async throws {
     let imageConfig = ImageConfig(aspectRatio: .landscape16x9)
@@ -403,7 +403,7 @@ struct GenerateContentIntegrationTests {
 
   @Test(arguments: [
     (InstanceConfig.googleAI_v1beta, ModelNames.gemini3_1_FlashImage),
-    (InstanceConfig.agentPlatform_v1beta_global, ModelNames.gemini3_1_FlashImage),
+    (InstanceConfig.enterprise_v1beta_global, ModelNames.gemini3_1_FlashImage),
   ])
   func generateImageWithCustomSize(_ config: InstanceConfig, modelName: String) async throws {
     let imageConfig = ImageConfig(
@@ -446,11 +446,11 @@ struct GenerateContentIntegrationTests {
   }
 
   @Test(arguments: [
-    (InstanceConfig.agentPlatform_v1beta, ModelNames.gemini2_5_FlashImage),
-    (InstanceConfig.agentPlatform_v1beta_global, ModelNames.gemini2_5_FlashImage),
+    (InstanceConfig.enterprise_v1beta, ModelNames.gemini2_5_FlashImage),
+    (InstanceConfig.enterprise_v1beta_global, ModelNames.gemini2_5_FlashImage),
     (InstanceConfig.googleAI_v1beta, ModelNames.gemini2_5_FlashImage),
     (InstanceConfig.googleAI_v1beta, ModelNames.gemini3_1_FlashImage),
-    (InstanceConfig.agentPlatform_v1beta_global, ModelNames.gemini3_1_FlashImage),
+    (InstanceConfig.enterprise_v1beta_global, ModelNames.gemini3_1_FlashImage),
   ])
   func generateContent_finishReason_imageSafety(_ config: InstanceConfig,
                                                 modelName: String) async throws {
@@ -466,7 +466,7 @@ struct GenerateContentIntegrationTests {
     do {
       let response = try await model.generateContent(prompt)
 
-      // agentPlatform gemini3_1_FlashImage doesn't throw.
+      // enterprise gemini3_1_FlashImage doesn't throw.
       let candidate = try #require(response.candidates.first)
       #expect(candidate.finishReason == .stop)
     } catch {
@@ -484,8 +484,8 @@ struct GenerateContentIntegrationTests {
   }
 
   @Test(arguments: [
-    (InstanceConfig.agentPlatform_v1beta, ModelNames.gemini2_5_FlashImage),
-    (InstanceConfig.agentPlatform_v1beta_global, ModelNames.gemini2_5_FlashImage),
+    (InstanceConfig.enterprise_v1beta, ModelNames.gemini2_5_FlashImage),
+    (InstanceConfig.enterprise_v1beta_global, ModelNames.gemini2_5_FlashImage),
     (InstanceConfig.googleAI_v1beta, ModelNames.gemini2_5_FlashImage),
     // Note: The following configs are commented out for easy one-off manual testing.
     // (InstanceConfig.googleAI_v1beta_staging, ModelNames.gemini2_5_FlashImage)
@@ -562,7 +562,7 @@ struct GenerateContentIntegrationTests {
   }
 
   // TODO: b/524685673 - Re-enable googleAI_v1beta once backend bug is fixed.
-  @Test("generateContent with URL Context", arguments: [InstanceConfig.agentPlatform_v1beta_global])
+  @Test("generateContent with URL Context", arguments: [InstanceConfig.enterprise_v1beta_global])
   func generateContent_withURLContext_succeeds(_ config: InstanceConfig) async throws {
     let model = FirebaseAI.componentInstance(config).generativeModel(
       modelName: ModelNames.gemini3_1_FlashLite,
@@ -612,14 +612,14 @@ struct GenerateContentIntegrationTests {
   // MARK: Streaming Tests
 
   @Test(arguments: [
-    (InstanceConfig.agentPlatform_v1beta, ModelNames.gemini2_5_FlashLite),
-    (InstanceConfig.agentPlatform_v1beta_global, ModelNames.gemini3_1_FlashLite),
-    (InstanceConfig.agentPlatform_v1beta_global_appCheckLimitedUse, ModelNames.gemini3_1_FlashLite),
+    (InstanceConfig.enterprise_v1beta, ModelNames.gemini2_5_FlashLite),
+    (InstanceConfig.enterprise_v1beta_global, ModelNames.gemini3_1_FlashLite),
+    (InstanceConfig.enterprise_v1beta_global_appCheckLimitedUse, ModelNames.gemini3_1_FlashLite),
     (InstanceConfig.googleAI_v1beta, ModelNames.gemini2_5_FlashLite),
     (InstanceConfig.googleAI_v1beta_appCheckLimitedUse, ModelNames.gemini2_5_FlashLite),
     (InstanceConfig.googleAI_v1beta, ModelNames.gemma4_31B),
     // Note: The following configs are commented out for easy one-off manual testing.
-    // (InstanceConfig.agentPlatform_v1beta_staging, ModelNames.gemini2_5_FlashLite),
+    // (InstanceConfig.enterprise_v1beta_staging, ModelNames.gemini2_5_FlashLite),
     // (InstanceConfig.googleAI_v1beta_staging, ModelNames.gemini2_5_FlashLite),
     // (InstanceConfig.googleAI_v1beta_staging, ModelNames.gemma4_31B),
     // (InstanceConfig.googleAI_v1beta_freeTier_bypassProxy, ModelNames.gemini2_5_FlashLite),
@@ -693,8 +693,8 @@ struct GenerateContentIntegrationTests {
   }
 
   @Test(arguments: [
-    (InstanceConfig.agentPlatform_v1beta, ModelNames.gemini2_5_FlashImage),
-    (InstanceConfig.agentPlatform_v1beta_global, ModelNames.gemini2_5_FlashImage),
+    (InstanceConfig.enterprise_v1beta, ModelNames.gemini2_5_FlashImage),
+    (InstanceConfig.enterprise_v1beta_global, ModelNames.gemini2_5_FlashImage),
     (InstanceConfig.googleAI_v1beta, ModelNames.gemini2_5_FlashImage),
     // Note: The following configs are commented out for easy one-off manual testing.
     // (InstanceConfig.googleAI_v1beta_staging, ModelNames.gemini2_5_FlashImage)