+++ FirebaseAI/Tests/Unit/GenerativeModelVertexAITests.swift
@@ -1241,7 +1241,7 @@ final class GenerativeModelVertexAITests: XCTestCase {
     XCTFail("Should have caught an error.")
   }
 
-  func testGenerateContentStream_failure_agentPlatformInFirebaseAPINotEnabled() async throws {
+  func testGenerateContentStream_failure_enterpriseInFirebaseAPINotEnabled() async throws {
     let expectedStatusCode = 403
     MockURLProtocol
       .requestHandler = try GenerativeModelTestUtil.httpRequestHandler(