+++ FirebaseAI/Tests/Unit/TestUtilities/FirebaseAI+DefaultAPIConfig.swift
@@ -16,7 +16,7 @@
 
 extension FirebaseAI {
   static let defaultVertexAIAPIConfig = APIConfig(
-    service: .agentPlatform(endpoint: .firebaseProxyProd, location: "us-central1"),
+    service: .enterprise(endpoint: .firebaseProxyProd, location: "us-central1"),
     version: .v1beta
   )
 }